refactor : tidt up packages and files
