+++ src/main/java/kitchenpos/order/application/OrderService.java
@@ -3,11 +3,11 @@ package kitchenpos.order.application;
 import kitchenpos.menu.domain.MenuRepository;
 import kitchenpos.order.domain.OrderRepository;
 import kitchenpos.dao.OrderLineItemDao;
-import kitchenpos.table.dto.OrderTableRepository;
-import kitchenpos.domain.OrderRequest;
+import kitchenpos.table.domain.OrderTableRepository;
+import kitchenpos.order.dto.OrderRequest;
 import kitchenpos.domain.OrderLineItemRequest;
-import kitchenpos.domain.OrderStatus;
-import kitchenpos.domain.OrderTableRequest;
+import kitchenpos.order.domain.OrderStatus;
+import kitchenpos.table.dto.OrderTableRequest;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.util.CollectionUtils;
