+++ src/main/java/kitchenpos/order/domain/OrderRepository.java
@@ -1,10 +1,6 @@
 package kitchenpos.order.domain;
 
-import kitchenpos.domain.OrderRequest;
-import kitchenpos.order.domain.Order;
-
 import java.util.List;
-import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
