+++ src/main/java/kitchenpos/order/ui/OrderRestController.java
@@ -1,7 +1,7 @@
 package kitchenpos.order.ui;
 
 import kitchenpos.order.application.OrderService;
-import kitchenpos.domain.OrderRequest;
+import kitchenpos.order.dto.OrderRequest;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
