+++ src/test/java/kitchenpos/acceptance/TableAcceptanceTest.java
@@ -21,9 +21,9 @@ import io.restassured.response.Response;
 import kitchenpos.menu.dto.MenuRequest;
 import kitchenpos.menuGroup.dto.MenuGroupRequest;
 import kitchenpos.menu.dto.MenuProductRequest;
-import kitchenpos.domain.OrderRequest;
+import kitchenpos.order.dto.OrderRequest;
 import kitchenpos.domain.OrderLineItemRequest;
-import kitchenpos.domain.OrderTableRequest;
+import kitchenpos.table.dto.OrderTableRequest;
 import kitchenpos.product.dto.ProductRequest;
 import kitchenpos.tableGroup.dto.TableGroupRequest;
