+++ src/test/java/kitchenpos/ui/OrderRestControllerTest.java
@@ -17,7 +17,7 @@ import org.springframework.http.MediaType;
 
 import kitchenpos.order.application.OrderService;
 import kitchenpos.order.ui.OrderRestController;
-import kitchenpos.domain.OrderRequest;
+import kitchenpos.order.dto.OrderRequest;
 import kitchenpos.domain.OrderLineItemRequest;
 
 @WebMvcTest(OrderRestController.class)
